refactor: proptypes + named export
