+++ src/Home/Home.container.js
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { Context } from 'context/Context';
-import Home from './Home';
+import { Home } from './Home';
 
 const EnhancedHome = props => {
   const { context } = useContext(Context);
