+++ src/Home/index.js
@@ -1,1 +1,2 @@
-export { default } from './Home';
+export { Home } from './Home';
+export { default } from './Home.container';
