+++ src/layout/Header/Header.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Header from './Header';
+import { HeaderApp } from './Header';
 
 it('1. Renders Header component without crashing', () => {
-  shallow(<Header />);
+  shallow(<HeaderApp />);
 });
