+++ src/layout/Header/index.js
@@ -1,1 +1,2 @@
-export { default } from './Header';
+export { HeaderApp } from './Header';
+export { default } from './Header.container';
