+++ src/layout/Menu/Menu.container.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Menu from './Menu';
+import { Menu } from './Menu';
 
 const EnhancedMenu = props => {
   return <Menu {...props} />;
