+++ src/layout/NotFound/NotFound.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './NotFound.scss';
 
-export default () => (
+export const NotFound = () => (
   <>
     <div className="container af-container--notfound">
       <h1 className="af-notfound__title">
