+++ src/layout/NotFound/index.js
@@ -1,1 +1,1 @@
-export { default } from './NotFound';
+export { NotFound } from './NotFound';
