+++ src/layout/ServerError/ServerError.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import './Error.scss';
+import './ServerError.scss';
 
-export default () => (
+export const ServerError = () => (
   <>
     <div className="container af-container--notfound">
       <h1 className="af-notfound__title">
