+++ src/layout/ServerError/ServerError.spec.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import { ServerError } from './ServerError';
+
+it('1. Renders ServerError component without crashing', () => {
+  shallow(<ServerError />);
+});
