+++ src/layout/ServerError/index.js
@@ -1,0 +1,1 @@
+export { ServerError } from './ServerError';
