+++ src/layout/Titlebar/Titlebar.js
@@ -1,10 +1,17 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import './Titlebar.scss';
 
-export default ({ title }) => (
+export const Titlebar = ({ title }) => (
   <div className="af-title-bar">
     <div className="af-title-bar__wrapper container-fluid container">
       <h1 className="af-title-bar__title">{title}</h1>
     </div>
   </div>
 );
+
+const propTypes = {
+  title: PropTypes.string.isRequired,
+};
+
+Titlebar.propTypes = propTypes;
