+++ src/layout/Titlebar/index.js
@@ -1,1 +1,2 @@
-export { default } from './Titlebar';
+export { Titlebar } from './Titlebar';
+export { default } from './Titlebar.container';
